refactor: share ecmascript export analysis
