+++ crates/monochange_ecmascript/readme.md
@@ -1,0 +1,5 @@
+# monochange_ecmascript
+
+`monochange_ecmascript` provides shared JavaScript and TypeScript semantic-analysis helpers for `monochange` ecosystem adapters.
+
+Use this crate when multiple adapters need the same ECMAScript module export extraction while keeping ecosystem-specific manifest logic in their own crates.
