+++ crates/monochange_npm/src/analysis.rs
@@ -1,6 +1,5 @@
 use std::collections::BTreeMap;
 use std::path::Path;
-use std::path::PathBuf;
 
 use monochange_core::AnalyzedFileChange;
 use monochange_core::DetectionLevel;
@@ -9,21 +8,13 @@ use monochange_core::MonochangeResult;
 use monochange_core::PackageAnalysisContext;
 use monochange_core::PackageAnalysisResult;
 use monochange_core::PackageRecord;
-use monochange_core::PackageSnapshot;
-use monochange_core::PackageSnapshotFile;
 use monochange_core::SemanticAnalyzer;
 use monochange_core::SemanticChange;
 use monochange_core::SemanticChangeCategory;
 use monochange_core::SemanticChangeKind;
-use oxc_allocator::Allocator;
-use oxc_ast::ast::Declaration;
-use oxc_ast::ast::ExportDefaultDeclarationKind;
-use oxc_ast::ast::ImportOrExportKind;
-use oxc_ast::ast::ModuleDeclaration;
-use oxc_ast::ast::ModuleExportName;
-use oxc_ast::ast::TSModuleDeclarationName;
-use oxc_parser::Parser;
-use oxc_span::SourceType;
+use monochange_ecmascript::EcmascriptExportConfig;
+use monochange_ecmascript::diff_public_symbols;
+use monochange_ecmascript::snapshot_exported_symbols;
 use serde_json::Value;
 
 use crate::PACKAGE_JSON_FILE;
@@ -55,10 +46,16 @@ impl SemanticAnalyzer for NpmSemanticAnalyzer {
 		let mut warnings = Vec::new();
 
 		if context.detection_level != DetectionLevel::Basic {
-			let before_symbols =
-				snapshot_exported_symbols(context.before_snapshot, context.changed_files);
-			let after_symbols =
-				snapshot_exported_symbols(context.after_snapshot, context.changed_files);
+			let before_symbols = snapshot_exported_symbols(
+				context.before_snapshot,
+				context.changed_files,
+				&NPM_ECMASCRIPT_EXPORT_CONFIG,
+			);
+			let after_symbols = snapshot_exported_symbols(
+				context.after_snapshot,
+				context.changed_files,
+				&NPM_ECMASCRIPT_EXPORT_CONFIG,
+			);
 			semantic_changes.extend(diff_public_symbols(&before_symbols, &after_symbols));
 		}
 
@@ -112,627 +109,14 @@ fn display_package_id(package: &PackageRecord) -> String {
 		.unwrap_or_else(|| package.id.clone())
 }
 
-#[derive(Debug, Clone, Eq, Ord, PartialEq, PartialOrd)]
-struct PublicSymbol {
-	item_kind: String,
-	item_path: String,
-	signature: String,
-	file_path: PathBuf,
-}
-
-fn snapshot_exported_symbols(
-	snapshot: Option<&PackageSnapshot>,
-	changed_files: &[AnalyzedFileChange],
-) -> BTreeMap<(String, String), PublicSymbol> {
-	let mut symbols = BTreeMap::new();
-
-	if let Some(snapshot) = snapshot {
-		for file in &snapshot.files {
-			if !is_source_file(&file.path) {
-				continue;
-			}
-
-			for symbol in collect_public_symbols(file) {
-				symbols.insert((symbol.item_kind.clone(), symbol.item_path.clone()), symbol);
-			}
-		}
-
-		return symbols;
-	}
-
-	for change in changed_files {
-		let Some(contents) = change
-			.after_contents
-			.as_deref()
-			.or(change.before_contents.as_deref())
-		else {
-			continue;
-		};
-		if !is_source_file(&change.package_path) {
-			continue;
-		}
-
-		let file = PackageSnapshotFile {
-			path: change.package_path.clone(),
-			contents: contents.to_string(),
-		};
-		for symbol in collect_public_symbols(&file) {
-			symbols.insert((symbol.item_kind.clone(), symbol.item_path.clone()), symbol);
-		}
-	}
-
-	symbols
-}
-
-fn is_source_file(path: &Path) -> bool {
-	let Some(extension) = path.extension().and_then(|ext| ext.to_str()) else {
-		return false;
-	};
-	if !matches!(
-		extension,
-		"js" | "jsx" | "ts" | "tsx" | "mjs" | "cjs" | "mts" | "cts"
-	) {
-		return false;
-	}
-
-	!path.starts_with("dist") && !path.starts_with("build") && !path.starts_with("node_modules")
-}
-
-fn collect_public_symbols(file: &PackageSnapshotFile) -> Vec<PublicSymbol> {
-	parse_public_symbols(file).unwrap_or_else(|| collect_public_symbols_with_legacy_scanner(file))
-}
-
-fn parse_public_symbols(file: &PackageSnapshotFile) -> Option<Vec<PublicSymbol>> {
-	let source_type = SourceType::from_path(&file.path).ok()?;
-	let allocator = Allocator::default();
-	let parser_return = Parser::new(&allocator, &file.contents, source_type).parse();
-	let module_prefix = module_prefix_for_file(&file.path);
-	let mut symbols = Vec::new();
-	for statement in &parser_return.program.body {
-		let Some(declaration) = statement.as_module_declaration() else {
-			continue;
-		};
-		collect_public_symbols_from_module_declaration(
-			declaration,
-			&module_prefix,
-			&file.path,
-			&file.contents,
-			&mut symbols,
-		);
-	}
-
-	Some(symbols)
-}
-
-fn collect_public_symbols_from_module_declaration(
-	declaration: &ModuleDeclaration<'_>,
-	module_prefix: &[String],
-	file_path: &Path,
-	source_text: &str,
-	output: &mut Vec<PublicSymbol>,
-) {
-	match declaration {
-		ModuleDeclaration::ExportNamedDeclaration(export) => {
-			let signature = normalize_signature(export.span.source_text(source_text));
-			if let Some(declaration) = &export.declaration {
-				collect_public_symbols_from_declaration(
-					declaration,
-					module_prefix,
-					file_path,
-					&signature,
-					output,
-				);
-				return;
-			}
-
-			for specifier in &export.specifiers {
-				let item_kind = if matches!(export.export_kind, ImportOrExportKind::Type)
-					|| matches!(specifier.export_kind, ImportOrExportKind::Type)
-				{
-					"type_reexport"
-				} else {
-					"reexport"
-				};
-				push_symbol(
-					output,
-					item_kind,
-					module_prefix,
-					module_export_name(&specifier.exported),
-					&signature,
-					file_path,
-				);
-			}
-		}
-		ModuleDeclaration::ExportAllDeclaration(export) => {
-			let signature = normalize_signature(export.span.source_text(source_text));
-			if let Some(exported) = &export.exported {
-				push_symbol(
-					output,
-					"namespace_reexport",
-					module_prefix,
-					module_export_name(exported),
-					&signature,
-					file_path,
-				);
-				return;
-			}
-
-			push_symbol(
-				output,
-				"wildcard_reexport",
-				module_prefix,
-				export.source.value.to_string(),
-				&signature,
-				file_path,
-			);
-		}
-		ModuleDeclaration::ExportDefaultDeclaration(export) => {
-			let signature = normalize_signature(export.span.source_text(source_text));
-			collect_public_symbols_from_default_export(
-				&export.declaration,
-				module_prefix,
-				file_path,
-				&signature,
-				output,
-			);
-		}
-		_ => {}
-	}
-}
-
-fn collect_public_symbols_from_declaration(
-	declaration: &Declaration<'_>,
-	module_prefix: &[String],
-	file_path: &Path,
-	signature: &str,
-	output: &mut Vec<PublicSymbol>,
-) {
-	match declaration {
-		Declaration::FunctionDeclaration(function) => {
-			if let Some(identifier) = &function.id {
-				push_symbol(
-					output,
-					"function",
-					module_prefix,
-					identifier.name.to_string(),
-					signature,
-					file_path,
-				);
-			}
-		}
-		Declaration::ClassDeclaration(class) => {
-			if let Some(identifier) = &class.id {
-				push_symbol(
-					output,
-					"class",
-					module_prefix,
-					identifier.name.to_string(),
-					signature,
-					file_path,
-				);
-			}
-		}
-		Declaration::VariableDeclaration(variable) => {
-			let item_kind = variable_item_kind(variable.kind);
-			for declarator in &variable.declarations {
-				for identifier in declarator.id.get_binding_identifiers() {
-					push_symbol(
-						output,
-						item_kind,
-						module_prefix,
-						identifier.name.to_string(),
-						signature,
-						file_path,
-					);
-				}
-			}
-		}
-		Declaration::TSInterfaceDeclaration(interface) => {
-			push_symbol(
-				output,
-				"interface",
-				module_prefix,
-				interface.id.name.to_string(),
-				signature,
-				file_path,
-			);
-		}
-		Declaration::TSTypeAliasDeclaration(alias) => {
-			push_symbol(
-				output,
-				"type_alias",
-				module_prefix,
-				alias.id.name.to_string(),
-				signature,
-				file_path,
-			);
-		}
-		Declaration::TSEnumDeclaration(declaration) => {
-			push_symbol(
-				output,
-				"enum",
-				module_prefix,
-				declaration.id.name.to_string(),
-				signature,
-				file_path,
-			);
-		}
-		Declaration::TSModuleDeclaration(namespace) => {
-			push_symbol(
-				output,
-				"namespace",
-				module_prefix,
-				ts_module_name(&namespace.id),
-				signature,
-				file_path,
-			);
-		}
-		Declaration::TSGlobalDeclaration(_) | Declaration::TSImportEqualsDeclaration(_) => {}
-	}
-}
-
-fn collect_public_symbols_from_default_export(
-	declaration: &ExportDefaultDeclarationKind<'_>,
-	module_prefix: &[String],
-	file_path: &Path,
-	signature: &str,
-	output: &mut Vec<PublicSymbol>,
-) {
-	match declaration {
-		ExportDefaultDeclarationKind::FunctionDeclaration(function) => {
-			if let Some(identifier) = &function.id {
-				push_symbol(
-					output,
-					"function",
-					module_prefix,
-					identifier.name.to_string(),
-					signature,
-					file_path,
-				);
-			} else {
-				push_symbol(
-					output,
-					"default_export",
-					module_prefix,
-					"default".to_string(),
-					signature,
-					file_path,
-				);
-			}
-		}
-		ExportDefaultDeclarationKind::ClassDeclaration(class) => {
-			if let Some(identifier) = &class.id {
-				push_symbol(
-					output,
-					"class",
-					module_prefix,
-					identifier.name.to_string(),
-					signature,
-					file_path,
-				);
-			} else {
-				push_symbol(
-					output,
-					"default_export",
-					module_prefix,
-					"default".to_string(),
-					signature,
-					file_path,
-				);
-			}
-		}
-		ExportDefaultDeclarationKind::TSInterfaceDeclaration(interface) => {
-			push_symbol(
-				output,
-				"interface",
-				module_prefix,
-				interface.id.name.to_string(),
-				signature,
-				file_path,
-			);
-		}
-		_ => {
-			push_symbol(
-				output,
-				"default_export",
-				module_prefix,
-				"default".to_string(),
-				signature,
-				file_path,
-			);
-		}
-	}
-}
-
-fn module_export_name(name: &ModuleExportName<'_>) -> String {
-	name.to_string()
-}
-
-fn ts_module_name(name: &TSModuleDeclarationName<'_>) -> String {
-	name.to_string()
-}
-
-fn variable_item_kind(kind: oxc_ast::ast::VariableDeclarationKind) -> &'static str {
-	match kind {
-		oxc_ast::ast::VariableDeclarationKind::Const => "constant",
-		oxc_ast::ast::VariableDeclarationKind::Var
-		| oxc_ast::ast::VariableDeclarationKind::Let
-		| oxc_ast::ast::VariableDeclarationKind::Using
-		| oxc_ast::ast::VariableDeclarationKind::AwaitUsing => "variable",
-	}
-}
-
-fn collect_public_symbols_with_legacy_scanner(file: &PackageSnapshotFile) -> Vec<PublicSymbol> {
-	let module_prefix = module_prefix_for_file(&file.path);
-	let mut symbols = Vec::new();
-
-	for raw_line in file.contents.lines() {
-		let line = normalize_signature(raw_line);
-		if !line.starts_with("export ") || line == "export {}" || line == "export {};" {
-			continue;
-		}
-
-		if let Some(target) = parse_wildcard_reexport(&line) {
-			push_symbol(
-				&mut symbols,
-				"wildcard_reexport",
-				&module_prefix,
-				&target,
-				&line,
-				&file.path,
-			);
-			continue;
-		}
-
-		let named_exports = parse_named_exports(&line);
-		if !named_exports.is_empty() {
-			for (item_kind, item_name) in named_exports {
-				push_symbol(
-					&mut symbols,
-					&item_kind,
-					&module_prefix,
-					&item_name,
-					&line,
-					&file.path,
-				);
-			}
-			continue;
-		}
-
-		if let Some((item_kind, item_name)) = parse_declaration_export(&line) {
-			push_symbol(
-				&mut symbols,
-				item_kind,
-				&module_prefix,
-				&item_name,
-				&line,
-				&file.path,
-			);
-		}
-	}
-
-	symbols
-}
-
-fn normalize_signature(line: &str) -> String {
-	line.split_whitespace().collect::<Vec<_>>().join(" ")
-}
-
-fn parse_wildcard_reexport(line: &str) -> Option<String> {
-	let rest = line.strip_prefix("export ")?;
-	let target = rest.strip_prefix("* from ")?;
-	extract_quoted_text(target)
-}
-
-fn parse_named_exports(line: &str) -> Vec<(String, String)> {
-	let Some(mut rest) = line.strip_prefix("export ") else {
-		return Vec::new();
-	};
-	let mut item_kind = "reexport";
-	if let Some(stripped) = rest.strip_prefix("type ") {
-		rest = stripped;
-		item_kind = "type_reexport";
-	}
-	if !rest.starts_with('{') {
-		return Vec::new();
-	}
-	let brace_start = rest
-		.find('{')
-		.expect("named exports should include an opening brace after the prefix check");
-	let Some(brace_end) = rest[brace_start + 1..].find('}') else {
-		return Vec::new();
-	};
-	let inside = &rest[brace_start + 1..brace_start + 1 + brace_end];
-
-	inside
-		.split(',')
-		.filter_map(|entry| {
-			let entry = entry.trim();
-			if entry.is_empty() {
-				return None;
-			}
-			let exported_name = entry.split(" as ").last().unwrap_or(entry).trim();
-			(!exported_name.is_empty()).then(|| (item_kind.to_string(), exported_name.to_string()))
-		})
-		.collect()
-}
-
-fn parse_declaration_export(line: &str) -> Option<(&'static str, String)> {
-	let mut rest = line.strip_prefix("export ")?;
-	if let Some(stripped) = rest.strip_prefix("declare ") {
-		rest = stripped;
-	}
-	if let Some(stripped) = rest.strip_prefix("default ") {
-		rest = stripped;
-	}
-	if let Some(stripped) = rest.strip_prefix("async ") {
-		rest = stripped;
-	}
-	if let Some(stripped) = rest.strip_prefix("abstract ") {
-		rest = stripped;
-	}
-
-	for (prefix, item_kind) in [
-		("function ", "function"),
-		("class ", "class"),
-		("const ", "constant"),
-		("let ", "variable"),
-		("var ", "variable"),
-		("interface ", "interface"),
-		("type ", "type_alias"),
-		("enum ", "enum"),
-		("namespace ", "namespace"),
-	] {
-		if let Some(stripped) = rest.strip_prefix(prefix)
-			&& let Some(name) = take_identifier(stripped)
-		{
-			return Some((item_kind, name));
-		}
-	}
-
-	Some(("default_export", "default".to_string())).filter(|_| line.starts_with("export default "))
-}
-
-fn take_identifier(text: &str) -> Option<String> {
-	let identifier = text
-		.chars()
-		.take_while(|character| character.is_ascii_alphanumeric() || matches!(character, '_' | '$'))
-		.collect::<String>();
-	(!identifier.is_empty()).then_some(identifier)
-}
-
-fn extract_quoted_text(text: &str) -> Option<String> {
-	let quote = text
-		.chars()
-		.find(|character| matches!(character, '\'' | '"'))?;
-	let start = text.find(quote)? + 1;
-	let end = text[start..].find(quote)? + start;
-	Some(text[start..end].to_string())
-}
-
-fn module_prefix_for_file(path: &Path) -> Vec<String> {
-	let mut components = path
-		.parent()
-		.map(|parent| {
-			parent
-				.components()
-				.filter_map(|component| component.as_os_str().to_str())
-				.map(ToString::to_string)
-				.collect::<Vec<_>>()
-		})
-		.unwrap_or_default();
-
-	if components
-		.first()
-		.is_some_and(|component| component == "src" || component == "lib")
-	{
-		components.remove(0);
-	}
-
-	let stem = path
-		.file_stem()
-		.and_then(|stem| stem.to_str())
-		.map(|stem| stem.strip_suffix(".d").unwrap_or(stem))
-		.unwrap_or_default();
-	if !stem.is_empty() && stem != "index" {
-		components.push(stem.to_string());
-	}
-
-	components
-}
-
-fn push_symbol<S: Into<String>>(
-	output: &mut Vec<PublicSymbol>,
-	item_kind: &str,
-	module_prefix: &[String],
-	item_name: S,
-	signature: &str,
-	file_path: &Path,
-) {
-	let item_name = item_name.into();
-	let item_path = if module_prefix.is_empty() {
-		item_name.clone()
-	} else {
-		format!("{}::{item_name}", module_prefix.join("::"))
-	};
-
-	output.push(PublicSymbol {
-		item_kind: item_kind.to_string(),
-		item_path,
-		signature: signature.to_string(),
-		file_path: file_path.to_path_buf(),
-	});
-}
-
-fn diff_public_symbols(
-	before: &BTreeMap<(String, String), PublicSymbol>,
-	after: &BTreeMap<(String, String), PublicSymbol>,
-) -> Vec<SemanticChange> {
-	let mut changes = Vec::new();
-
-	for (key, after_symbol) in after {
-		match before.get(key) {
-			None => {
-				changes.push(build_symbol_change(
-					SemanticChangeKind::Added,
-					after_symbol,
-					None,
-					Some(after_symbol.signature.clone()),
-				));
-			}
-			Some(before_symbol) if before_symbol.signature != after_symbol.signature => {
-				changes.push(build_symbol_change(
-					SemanticChangeKind::Modified,
-					after_symbol,
-					Some(before_symbol.signature.clone()),
-					Some(after_symbol.signature.clone()),
-				));
-			}
-			Some(_) => {}
-		}
-	}
-
-	for (key, before_symbol) in before {
-		if after.contains_key(key) {
-			continue;
-		}
-
-		changes.push(build_symbol_change(
-			SemanticChangeKind::Removed,
-			before_symbol,
-			Some(before_symbol.signature.clone()),
-			None,
-		));
-	}
-
-	changes
-}
-
-fn build_symbol_change(
-	kind: SemanticChangeKind,
-	symbol: &PublicSymbol,
-	before_signature: Option<String>,
-	after_signature: Option<String>,
-) -> SemanticChange {
-	let verb = if kind == SemanticChangeKind::Added {
-		"added"
-	} else if kind == SemanticChangeKind::Removed {
-		"removed"
-	} else {
-		"modified"
-	};
-
-	SemanticChange {
-		category: SemanticChangeCategory::PublicApi,
-		kind,
-		item_kind: symbol.item_kind.clone(),
-		item_path: symbol.item_path.clone(),
-		summary: format!("{} `{}` {verb}", symbol.item_kind, symbol.item_path),
-		file_path: symbol.file_path.clone(),
-		before_signature,
-		after_signature,
-	}
-}
+const NPM_ECMASCRIPT_EXPORT_CONFIG: EcmascriptExportConfig = EcmascriptExportConfig {
+	source_extensions: &["js", "jsx", "ts", "tsx", "mjs", "cjs", "mts", "cts"],
+	module_roots_to_strip: &["src", "lib"],
+	ignored_module_stems: &["index"],
+	strip_declaration_stem_suffix: true,
+	legacy_supports_declare_prefix: true,
+	legacy_supports_namespace_exports: true,
+};
 
 #[derive(Debug, Clone, Eq, PartialEq, Ord, PartialOrd)]
 struct ManifestEntry {
@@ -1058,174 +442,11 @@ fn describe_json_value(value: &Value) -> String {
 
 #[cfg(test)]
 mod tests {
+	use std::path::PathBuf;
+
 	use monochange_core::FileChangeKind;
 
 	use super::*;
-
-	#[test]
-	fn module_prefix_for_nested_source_file_tracks_directory_components() {
-		assert_eq!(
-			module_prefix_for_file(Path::new("src/api/index.ts")),
-			vec!["api".to_string()]
-		);
-		assert_eq!(
-			module_prefix_for_file(Path::new("src/utils/format.ts")),
-			vec!["utils".to_string(), "format".to_string()]
-		);
-	}
-
-	#[test]
-	fn collect_public_symbols_finds_exported_declarations_and_reexports() {
-		let file = PackageSnapshotFile {
-			path: PathBuf::from("src/index.ts"),
-			contents: concat!(
-				"export function greet(name: string): string { return name; }\n",
-				"export const version = '1.0.0';\n",
-				"export { formatGreeting as format } from './format';\n",
-				"export * from './types';\n",
-			)
-			.to_string(),
-		};
-
-		let symbols = collect_public_symbols(&file);
-
-		assert!(symbols.iter().any(|symbol| symbol.item_path == "greet"));
-		assert!(symbols.iter().any(|symbol| symbol.item_path == "version"));
-		assert!(symbols.iter().any(|symbol| symbol.item_path == "format"));
-		assert!(symbols.iter().any(|symbol| {
-			symbol.item_kind == "wildcard_reexport" && symbol.item_path == "./types"
-		}));
-	}
-
-	#[test]
-	fn parser_backed_symbol_collection_handles_multiline_and_namespace_exports() {
-		let file = PackageSnapshotFile {
-			path: PathBuf::from("src/index.ts"),
-			contents: concat!(
-				"export {\n",
-				"  greet as renamedGreet,\n",
-				"  version,\n",
-				"} from './shared';\n",
-				"export * as toolkit from './toolkit';\n",
-				"export default function () { return 'ok'; }\n",
-			)
-			.to_string(),
-		};
-
-		let symbols = collect_public_symbols(&file);
-
-		assert!(
-			symbols
-				.iter()
-				.any(|symbol| symbol.item_path == "renamedGreet")
-		);
-		assert!(symbols.iter().any(|symbol| symbol.item_path == "version"));
-		assert!(symbols.iter().any(|symbol| {
-			symbol.item_kind == "namespace_reexport" && symbol.item_path == "toolkit"
-		}));
-		assert!(symbols.iter().any(|symbol| {
-			symbol.item_kind == "default_export" && symbol.item_path == "default"
-		}));
-	}
-
-	#[test]
-	fn parser_backed_symbol_collection_handles_type_reexports_and_ts_declarations() {
-		let file = PackageSnapshotFile {
-			path: PathBuf::from("src/index.ts"),
-			contents: concat!(
-				"const hidden = true;\n",
-				"import { hiddenImport } from './hidden';\n",
-				"export { type Foo as Bar } from './types';\n",
-				"export interface Greeting {}\n",
-				"export type GreetingValue = string;\n",
-				"export enum Mode { Light }\n",
-				"export namespace Toolkit {}\n",
-				"export import Legacy = Toolkit;\n",
-				"export default {};\n",
-			)
-			.to_string(),
-		};
-
-		let symbols = collect_public_symbols(&file);
-
-		assert!(
-			symbols
-				.iter()
-				.any(|symbol| { symbol.item_kind == "type_reexport" && symbol.item_path == "Bar" })
-		);
-		assert!(symbols.iter().any(|symbol| symbol.item_path == "Greeting"));
-		assert!(
-			symbols
-				.iter()
-				.any(|symbol| symbol.item_path == "GreetingValue")
-		);
-		assert!(symbols.iter().any(|symbol| symbol.item_path == "Mode"));
-		assert!(symbols.iter().any(|symbol| symbol.item_path == "Toolkit"));
-		assert!(symbols.iter().any(|symbol| {
-			symbol.item_kind == "default_export" && symbol.item_path == "default"
-		}));
-		assert!(!symbols.iter().any(|symbol| symbol.item_path == "hidden"));
-		assert!(!symbols.iter().any(|symbol| symbol.item_path == "Legacy"));
-	}
-
-	#[test]
-	fn parser_backed_symbol_collection_handles_named_default_exports() {
-		let default_function = PackageSnapshotFile {
-			path: PathBuf::from("src/index.ts"),
-			contents: "export default function namedDefault() { return 'ok'; }\n".to_string(),
-		};
-		let default_class = PackageSnapshotFile {
-			path: PathBuf::from("src/index.ts"),
-			contents: "export default class NamedGreeter {}\n".to_string(),
-		};
-		let anonymous_default_class = PackageSnapshotFile {
-			path: PathBuf::from("src/index.ts"),
-			contents: "export default class {}\n".to_string(),
-		};
-		let default_interface = PackageSnapshotFile {
-			path: PathBuf::from("src/index.ts"),
-			contents: "export default interface GreetingContract {}\n".to_string(),
-		};
-
-		let function_symbols = collect_public_symbols(&default_function);
-		let class_symbols = collect_public_symbols(&default_class);
-		let anonymous_class_symbols = collect_public_symbols(&anonymous_default_class);
-		let interface_symbols = collect_public_symbols(&default_interface);
-
-		assert!(function_symbols.iter().any(|symbol| {
-			symbol.item_kind == "function" && symbol.item_path == "namedDefault"
-		}));
-		assert!(
-			class_symbols.iter().any(|symbol| {
-				symbol.item_kind == "class" && symbol.item_path == "NamedGreeter"
-			})
-		);
-		assert!(anonymous_class_symbols.iter().any(|symbol| {
-			symbol.item_kind == "default_export" && symbol.item_path == "default"
-		}));
-		assert!(interface_symbols.iter().any(|symbol| {
-			symbol.item_kind == "interface" && symbol.item_path == "GreetingContract"
-		}));
-	}
-
-	#[test]
-	fn variable_item_kind_maps_non_const_bindings_to_variable() {
-		assert_eq!(
-			variable_item_kind(oxc_ast::ast::VariableDeclarationKind::Let),
-			"variable"
-		);
-	}
-
-	#[test]
-	fn legacy_scanner_collects_single_line_exports() {
-		let file = PackageSnapshotFile {
-			path: PathBuf::from("src/index.ts"),
-			contents: "export const version = '1.0.0';\n".to_string(),
-		};
-
-		let symbols = collect_public_symbols_with_legacy_scanner(&file);
-		assert!(symbols.iter().any(|symbol| symbol.item_path == "version"));
-	}
 
 	#[test]
 	fn analyze_manifest_change_reports_export_dependency_and_metadata_diffs() {
@@ -1302,65 +523,6 @@ mod tests {
 	}
 
 	#[test]
-	fn snapshot_and_symbol_helpers_cover_additional_export_forms() {
-		let changed_files = vec![
-			AnalyzedFileChange {
-				path: PathBuf::from("packages/web/src/index.ts"),
-				package_path: PathBuf::from("src/index.ts"),
-				kind: FileChangeKind::Modified,
-				before_contents: None,
-				after_contents: Some(
-					concat!(
-						"export declare async function greet(name: string): Promise<string> { return name; }\n",
-						"export default class Greeter {}\n",
-						"export abstract class BaseGreeter {}\n",
-						"export namespace Tools {}\n",
-						"export type { Foo as Bar } from './types';\n",
-					)
-					.to_string(),
-				),
-			},
-			AnalyzedFileChange {
-				path: PathBuf::from("packages/web/README.md"),
-				package_path: PathBuf::from("README.md"),
-				kind: FileChangeKind::Modified,
-				before_contents: None,
-				after_contents: Some("ignored".to_string()),
-			},
-			AnalyzedFileChange {
-				path: PathBuf::from("packages/web/src/legacy.ts"),
-				package_path: PathBuf::from("src/legacy.ts"),
-				kind: FileChangeKind::Modified,
-				before_contents: Some("export const previous = true;".to_string()),
-				after_contents: None,
-			},
-		];
-
-		let symbols = snapshot_exported_symbols(None, &changed_files);
-
-		for expected in [
-			"greet",
-			"Greeter",
-			"BaseGreeter",
-			"Tools",
-			"Bar",
-			"legacy::previous",
-		] {
-			assert!(
-				symbols
-					.iter()
-					.any(|((_, item_path), _)| item_path == expected)
-			);
-		}
-		assert!(is_source_file(Path::new("src/index.cts")));
-		assert!(!is_source_file(Path::new("build/index.ts")));
-		assert_eq!(
-			module_prefix_for_file(Path::new("lib/utils/index.d.ts")),
-			vec!["utils".to_string()]
-		);
-	}
-
-	#[test]
 	fn manifest_helpers_cover_parse_failures_removed_entries_and_scalar_bins() {
 		let mut warnings = Vec::new();
 		assert!(parse_manifest(Some("{"), Path::new("package.json"), &mut warnings).is_none());
@@ -1407,88 +569,6 @@ mod tests {
 			change.item_path == "script.build" && change.kind == SemanticChangeKind::Removed
 		}));
 
-		assert_eq!(describe_json_value(&serde_json::json!(null)), "null");
-		assert_eq!(describe_json_value(&serde_json::json!(true)), "true");
-		assert_eq!(describe_json_value(&serde_json::json!(3)), "3");
-		assert!(describe_json_value(&serde_json::json!(["a", "b"])).contains("a, b"));
-		assert!(describe_json_value(&serde_json::json!({"b": 2, "a": 1})).contains("a=1"));
-	}
-
-	#[test]
-	fn parser_diff_and_export_helpers_cover_remaining_npm_branches() {
-		let skipped_symbols = snapshot_exported_symbols(
-			None,
-			&[
-				AnalyzedFileChange {
-					path: PathBuf::from("packages/web/src/empty.ts"),
-					package_path: PathBuf::from("src/empty.ts"),
-					kind: FileChangeKind::Modified,
-					before_contents: None,
-					after_contents: None,
-				},
-				AnalyzedFileChange {
-					path: PathBuf::from("packages/web/README.md"),
-					package_path: PathBuf::from("README.md"),
-					kind: FileChangeKind::Modified,
-					before_contents: None,
-					after_contents: Some("ignored".to_string()),
-				},
-			],
-		);
-		assert!(skipped_symbols.is_empty());
-		assert!(!is_source_file(Path::new("src/index")));
-
-		assert!(parse_named_exports("const nope = true;").is_empty());
-		assert_eq!(
-			parse_named_exports("export type { Foo as Bar, , Baz } from './types';"),
-			vec![
-				("type_reexport".to_string(), "Bar".to_string()),
-				("type_reexport".to_string(), "Baz".to_string()),
-			]
-		);
-		assert!(parse_named_exports("export { Foo as Bar from './types';").is_empty());
-		assert_eq!(
-			parse_declaration_export("export default {};"),
-			Some(("default_export", "default".to_string()))
-		);
-
-		let before = BTreeMap::from([
-			(
-				("function".to_string(), "greet".to_string()),
-				PublicSymbol {
-					item_kind: "function".to_string(),
-					item_path: "greet".to_string(),
-					signature: "export function greet() {}".to_string(),
-					file_path: PathBuf::from("src/index.ts"),
-				},
-			),
-			(
-				("class".to_string(), "Greeter".to_string()),
-				PublicSymbol {
-					item_kind: "class".to_string(),
-					item_path: "Greeter".to_string(),
-					signature: "export class Greeter {}".to_string(),
-					file_path: PathBuf::from("src/index.ts"),
-				},
-			),
-		]);
-		let after = BTreeMap::from([(
-			("function".to_string(), "greet".to_string()),
-			PublicSymbol {
-				item_kind: "function".to_string(),
-				item_path: "greet".to_string(),
-				signature: "export function greet() {}".to_string(),
-				file_path: PathBuf::from("src/index.ts"),
-			},
-		)]);
-		let changes = diff_public_symbols(&before, &after);
-		assert_eq!(changes.len(), 1);
-		let change = changes
-			.first()
-			.unwrap_or_else(|| panic!("expected one removed change"));
-		assert_eq!(change.kind, SemanticChangeKind::Removed);
-		assert!(change.summary.contains("removed"));
-
 		let exports = extract_public_exports(
 			&serde_json::json!({
 				"exports": {".": "./dist/index.js", "./cli": "./dist/cli.js"},
@@ -1499,5 +579,11 @@ mod tests {
 		assert!(exports.contains_key("."));
 		assert!(exports.contains_key("./cli"));
 		assert!(!exports.contains_key("pkg"));
+
+		assert_eq!(describe_json_value(&serde_json::json!(null)), "null");
+		assert_eq!(describe_json_value(&serde_json::json!(true)), "true");
+		assert_eq!(describe_json_value(&serde_json::json!(3)), "3");
+		assert!(describe_json_value(&serde_json::json!(["a", "b"])).contains("a, b"));
+		assert!(describe_json_value(&serde_json::json!({"b": 2, "a": 1})).contains("a=1"));
 	}
 }
